refactor: assert 경로 설정
