+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export { getStoredData, setStoredData, clearStoredData } from './userStorage';
 export { default as getRelativeTime } from './getRelativeTime';
 export { default as getRandomItem } from './getRandomItem';
+export { default as assert } from './assert';
 export * from './validations';
